修改

+++ mktcenter-serviceimpl/src/main/java/com/bizvane/mktcenterserviceimpl/service/impl/ActivityRegisterServiceImpl.java
@@ -268,7 +268,7 @@ public class ActivityRegisterServiceImpl implements ActivityRegisterService {
            pageVo.setPageNum(1);
            pageVo.setPageSize(10000);
            membersInfoSearchVo.setCardStatus(1);
-           membersInfoSearchVo.setBrandId(activityVO.getSysBrandId().intValue());
+           membersInfoSearchVo.setBrandId(activityVO.getSysBrandId());
            ResponseData<com.bizvane.utils.responseinfo.PageInfo<MemberInfoVo>> memberInfoVoPage = membersAdvancedSearchApiService.search(membersInfoSearchVo,pageVo);
           //循环分页条件查询会员信息发送短信信息
            for (int a =1;i<=memberInfoVoPage.getData().getPages();a++){
